# proyecto-dws-grupo2/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: TitanGym API
  description: Documentación oficial de la API REST del gimnasio TitanGym
  version: "1.0"
servers:
- url: https://localhost:8443
  description: Generated server url
security:
- bearerAuth: []
...
paths:
  /api/v1/users/{id}:
    get:
      tags:
      - Usuarios
      summary: Obtener usuario por ID
      description: Busca la información de un miembro específico mediante su identificador.
      operationId: getUserById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: No existe ningún usuario con ese ID
          content: {}
        "200":
          description: Usuario encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
    put:
      tags:
      - Usuarios
      summary: Actualizar un usuario
      description: "Modifica los datos de un miembro existente. Si se envía una contraseñ\
        a, será encriptada."
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequestDTO"
        required: true
      responses:
        "404":
          description: Usuario no encontrado
          content: {}
        "200":
          description: Usuario actualizado correctamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "400":
          description: Datos de actualización inválidos
    delete:
      tags:
      - Usuarios
      summary: Eliminar un usuario
      description: Borra permanentemente la cuenta de un miembro.
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Usuario no encontrado
          content: {}
        "200":
          description: Usuario eliminado con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
...
  /api/v1/reviews/{id}:
    get:
      tags:
      - Reseñas
      summary: Obtener reseña por ID
      description: Busca una valoración específica por su identificador único.
      operationId: getReview
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: No existe ninguna reseña con ese ID
          content: {}
        "200":
          description: Reseña encontrada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewResponseDTO"
    put:
      tags:
      - Reseñas
      summary: Actualizar una reseña
      description: Modifica el contenido o la nota de una reseña existente.
      operationId: replaceReview
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequestDTO"
        required: true
      responses:
        "404":
          description: Reseña no encontrada
          content: {}
        "200":
          description: Reseña actualizada correctamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewResponseDTO"
    delete:
      tags:
      - Reseñas
      summary: Eliminar una reseña
      description: Borra permanentemente un comentario. Requiere permisos de administrador
        o del autor.
      operationId: deleteReview
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Reseña no encontrada
          content: {}
        "200":
          description: Reseña eliminada con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewResponseDTO"
...
  /api/v1/users/{userId}/image:
    post:
      tags:
      - Imágenes
      summary: Subir foto de perfil de usuario
      description: Sube una imagen y la asocia como foto de perfil de un usuario existente.
      operationId: uploadProfileImage
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "404":
          description: Usuario no encontrado
          content: {}
        "201":
          description: Imagen subida y asociada con éxito
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: El archivo enviado está vacío o es inválido
...
  /api/v1/services/{serviceId}/image:
    post:
      tags:
      - Imágenes
      summary: Subir imagen de un servicio
      description: Sube una imagen representativa para un servicio o suplemento.
      operationId: uploadServiceImage
      parameters:
      - name: serviceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "404":
          description: Servicio no encontrado
          content: {}
        "201":
          description: Imagen subida y asociada con éxito
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/reviews/:
    get:
      tags:
      - Reseñas
      summary: Obtener todas las reseñas
      description: Devuelve una página con todas las valoraciones de los clientes.
      operationId: getReviews
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
        "200":
          description: Lista de reseñas recuperada con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelReviewResponseDTO"
    post:
      tags:
      - Reseñas
      summary: Publicar una nueva reseña
      description: Crea un comentario y una puntuación asociada a una clase y un usuario.
      operationId: createReview
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "201":
          description: Reseña publicada con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewResponseDTO"
        "400":
          description: Datos de la reseña inválidos
...
  /api/v1/classes/{classId}/image:
    post:
      tags:
      - Imágenes
      summary: Subir imagen de una clase
      description: Sube una imagen representativa para una clase o actividad del gimnasio.
      operationId: uploadClassImage
      parameters:
      - name: classId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "404":
          description: Clase no encontrada
          content: {}
        "201":
          description: Imagen subida y asociada con éxito
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/images/{id}/media:
    get:
      tags:
      - Imágenes
      summary: Descargar una imagen
      description: Descarga los datos binarios de una imagen a partir de su ID.
      operationId: downloadImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Imagen no encontrada
          content: {}
        "200":
          description: Imagen descargada con éxito
          content:
            '*/*':
              schema:
                type: string
                format: byte
  /api/v1/classes/{id}/reviews:
    get:
      tags:
      - Clases
      summary: Obtener reseñas de una clase
      description: Recupera todas las valoraciones y comentarios asociados a una clase
        concreta.
      operationId: getClassReviews
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Clase no encontrada
          content: {}
        "200":
          description: Lista de reseñas recuperada con éxito
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    UserRequestDTO:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
    ImageResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        url:
          type: string
    UserResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        profileImage:
          $ref: "#/components/schemas/ImageResponseDTO"
...
    ReviewRequestDTO:
      type: object
      properties:
        about:
          type: string
        rating:
          type: integer
          format: int32
        comment:
          type: string
        userId:
          type: integer
          format: int64
        classEntityId:
          type: integer
          format: int64
    ReviewResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        about:
          type: string
        rating:
          type: integer
          format: int32
        comment:
          type: string
        user:
          $ref: "#/components/schemas/UserResponseDTO"
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
...
    PagedModelReviewResponseDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ReviewResponseDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT